Refactor ReviewForm and LocationForm
Move duplicated and nonessential code into constants and utils files
