+++ src/components/form/LocationForm.js
@@ -4,99 +4,36 @@ import { useDispatch, useSelector } from 'react-redux'
 import { Link, useLocation } from 'react-router-dom'
 import styled from 'styled-components/macro'
 
+import {
+  INITIAL_LOCATION_VALUES,
+  MONTH_OPTIONS,
+  PROPERTY_ACCESS_OPTIONS,
+} from '../../constants/form'
 import { saveFormValues, submitLocation } from '../../redux/locationSlice'
 import { pathWithCurrentView } from '../../utils/appUrl'
+import {
+  formToLocation,
+  formToReview,
+  isEmptyReview,
+  validateLocation,
+  validateLocationStep,
+  validateReviewStep,
+} from '../../utils/form'
 import { useAppHistory } from '../../utils/useAppHistory'
 import { useIsDesktop } from '../../utils/useBreakpoint'
 import Button from '../ui/Button'
 import IconBesideText from '../ui/IconBesideText'
 import Label from '../ui/Label'
 import LoadingIndicator from '../ui/LoadingIndicator'
-import FormikAllSteps from './FormikAllSteps'
-import { FormikStepper, ProgressButtons, Step } from './FormikStepper'
+import { ProgressButtons, Step } from './FormikStepper'
 import { Checkbox, Select, Textarea } from './FormikWrappers'
-import {
-  formToReview,
-  INITIAL_REVIEW_VALUES,
-  isEmptyReview,
-  ReviewPhotoStep,
-  ReviewStep,
-  validateReview,
-  validateReviewStep,
-} from './ReviewForm'
+import { FormWrapper } from './FormLayout'
+import { ReviewPhotoStep, ReviewStep } from './ReviewForm'
 import TypesSelect from './TypesSelect'
-import { useInvisibleRecaptcha } from './useInvisibleRecaptcha'
 
 const StyledPositionFieldLink = styled(Link)`
   color: ${({ theme }) => theme.orange};
   text-decoration: none;
-`
-
-const PROPERTY_ACCESS_LABELS = [
-  'Source is on my property',
-  'I have permission from the owner to add the source',
-  'Source is on public land',
-  'Source is on private property but overhangs public land',
-  'Source is on private property (ask before you pick)',
-]
-
-// TODO: move this to commmon utils constants
-export const MONTH_LABELS = [
-  'January',
-  'February',
-  'March',
-  'April',
-  'May',
-  'June',
-  'July',
-  'August',
-  'September',
-  'October',
-  'November',
-  'December',
-]
-
-const labelsToOptions = (labels) =>
-  labels.map((label, index) => ({
-    label,
-    value: index,
-  }))
-
-const PROPERTY_ACCESS_OPTIONS = labelsToOptions(PROPERTY_ACCESS_LABELS)
-
-const MONTH_OPTIONS = labelsToOptions(MONTH_LABELS)
-
-const INITIAL_LOCATION_VALUES = {
-  types: [],
-  description: '',
-  access: null,
-  season_start: null,
-  season_stop: null,
-  ...INITIAL_REVIEW_VALUES,
-}
-
-const StyledLocationForm = styled.div`
-  box-sizing: border-box;
-  width: 100%;
-  padding: 0 10px;
-  overflow: auto;
-
-  @media ${({ theme }) => theme.device.desktop} {
-    height: 100%;
-  }
-
-  @media ${({ theme }) => theme.device.mobile} {
-    padding: 8px 27px 20px;
-    margin-top: 80px;
-
-    textarea {
-      height: 100px;
-
-      @media (max-device-height: 600px) {
-        height: 50px;
-      }
-    }
-  }
 `
 
 const CheckboxLabel = styled.label`
@@ -195,58 +132,6 @@ const LocationStep = ({ lat, lng, isDesktop, editingId, isLoading }) => (
   </>
 )
 
-const validateLocationStep = ({ types }) => {
-  const errors = {}
-  if (types.length === 0) {
-    errors.types = true
-  }
-  return errors
-}
-
-const validateLocation = ({ review, ...location }) => {
-  const errors = validateLocationStep(location)
-
-  if (!isEmptyReview(review)) {
-    Object.assign(errors, validateReview(review))
-  }
-
-  return errors
-}
-
-const formToLocation = ({
-  types,
-  description,
-  season_start,
-  season_stop,
-  access,
-  unverified,
-}) => ({
-  type_ids: types.map(({ value }) => value),
-  description,
-  season_start: season_start?.value ?? null,
-  season_stop: season_stop?.value ?? null,
-  access: access?.value ?? null,
-  unverified,
-})
-
-export const locationToForm = (
-  { type_ids, description, season_start, season_stop, access, unverified },
-  typesAccess,
-) => ({
-  types: type_ids?.map((id) => {
-    const type = typesAccess.getType(id)
-    return {
-      ...type,
-      value: id,
-    }
-  }),
-  description,
-  season_start: MONTH_OPTIONS[season_start],
-  season_stop: MONTH_OPTIONS[season_stop],
-  access: PROPERTY_ACCESS_OPTIONS[access],
-  unverified,
-})
-
 export const LocationForm = ({ editingId, initialValues, stepped }) => {
   const reduxFormValues = useSelector((state) => state.location.form)
   const mergedInitialValues = {
@@ -261,7 +146,6 @@ export const LocationForm = ({ editingId, initialValues, stepped }) => {
   const dispatch = useDispatch()
 
   const { position, isLoading } = useSelector((state) => state.location)
-  const isLoggedIn = useSelector((state) => !!state.auth.user)
 
   const formikSteps = [
     <Step key={1} label="Step 1" validate={validateLocationStep}>
@@ -312,44 +196,33 @@ export const LocationForm = ({ editingId, initialValues, stepped }) => {
       }
     })
   }
-
-  const { Recaptcha, handlePresubmit } = useInvisibleRecaptcha(handleSubmit)
-
-  const StepDisplay = stepped ? FormikStepper : FormikAllSteps
+  const handleCancel = (e) => {
+    e.stopPropagation()
+    if (editingId) {
+      history.push(`/locations/${editingId}`)
+    } else {
+      history.push(state?.fromPage ?? '/map')
+    }
+  }
 
   return (
-    <StyledLocationForm>
-      <StepDisplay
-        validateOnChange={false}
-        validate={validateLocation}
-        initialValues={mergedInitialValues}
-        validateOnMount
-        onSubmit={isLoggedIn ? handleSubmit : handlePresubmit}
-        renderButtons={({ isSubmitting, isValid }) => (
-          <ProgressButtons>
-            <Button
-              secondary
-              type="button"
-              onClick={(e) => {
-                e.stopPropagation()
-                if (editingId) {
-                  history.push(`/locations/${editingId}`)
-                } else {
-                  history.push(state?.fromPage ?? '/map')
-                }
-              }}
-            >
-              Cancel
-            </Button>
-            <Button disabled={isSubmitting || !isValid} type="submit">
-              {isSubmitting ? 'Submitting' : 'Submit'}
-            </Button>
-          </ProgressButtons>
-        )}
-      >
-        {formikSteps}
-        {!isLoggedIn && <Recaptcha />}
-      </StepDisplay>
-    </StyledLocationForm>
+    <FormWrapper
+      validate={validateLocation}
+      initialValues={mergedInitialValues}
+      onSubmit={handleSubmit}
+      stepped={stepped}
+      renderButtons={({ isSubmitting, isValid }) => (
+        <ProgressButtons>
+          <Button secondary type="button" onClick={handleCancel}>
+            Cancel
+          </Button>
+          <Button disabled={isSubmitting || !isValid} type="submit">
+            {isSubmitting ? 'Submitting' : 'Submit'}
+          </Button>
+        </ProgressButtons>
+      )}
+    >
+      {formikSteps}
+    </FormWrapper>
   )
 }
