+++ src/components/form/ReviewForm.js
@@ -2,19 +2,19 @@ import { darken } from 'polished'
 import { useDispatch, useSelector } from 'react-redux'
 import styled from 'styled-components/macro'
 
+import { FRUITING_OPTIONS, INITIAL_REVIEW_VALUES } from '../../constants/form'
 import {
   deleteLocationReview,
   submitLocationReview,
 } from '../../redux/locationSlice'
+import { formToReview, validateReview } from '../../utils/form'
 import { useAppHistory } from '../../utils/useAppHistory'
 import { FormRatingWrapper } from '../auth/AuthWrappers'
-import { isEveryPhotoUploaded } from '../photo/PhotoUploader'
 import Button from '../ui/Button'
 import LabeledRow from '../ui/LabeledRow'
 import { Optional } from '../ui/LabelTag'
 import SectionHeading from '../ui/SectionHeading'
-import FormikAllSteps from './FormikAllSteps'
-import { FormikStepper, ProgressButtons, Step } from './FormikStepper'
+import { ProgressButtons } from './FormikStepper'
 import {
   DateInput,
   PhotoUploader,
@@ -22,115 +22,7 @@ import {
   Select,
   Textarea,
 } from './FormikWrappers'
-import { useInvisibleRecaptcha } from './useInvisibleRecaptcha'
-
-/**
- * Initial values of review form fields.
- *
- * These represent the value that each field resets to.
- */
-export const INITIAL_REVIEW_VALUES = {
-  review: {
-    photos: [],
-    comment: '',
-    observed_on: '',
-    fruiting: null,
-    quality_rating: '0',
-    yield_rating: '0',
-  },
-}
-
-/* TODO: Create a generic utility to build form options */
-export const FRUITING_OPTIONS = ['Flowers', 'Unripe fruit', 'Ripe fruit'].map(
-  (name, idx) => ({
-    label: name,
-    value: idx,
-  }),
-)
-
-export const isEmptyReview = (review) => {
-  if (!review) {
-    return true
-  }
-
-  const r = formToReview(review)
-  return (
-    !r.comment &&
-    r.quality_rating === null &&
-    r.fruiting === null &&
-    r.yield_rating === null &&
-    r.photo_ids.length === 0
-  )
-}
-
-export const validateReviewStep = (review) => {
-  const r = formToReview(review)
-  if (r.fruiting !== null && !r.observed_on) {
-    return {
-      review: { observed_on: true },
-    }
-  }
-
-  return null
-}
-
-export const validatePhotoStep = (review) => {
-  if (isEmptyReview(review)) {
-    return {
-      review: { comment: true },
-    }
-  }
-
-  if (!isEveryPhotoUploaded(review.photos)) {
-    return {
-      review: { photos: true },
-    }
-  }
-
-  return null
-}
-
-export const validateReview = (review) => ({
-  ...validatePhotoStep(review),
-  ...validateReviewStep(review),
-})
-
-export const formToReview = (review) => {
-  const formattedReview = {
-    ...review,
-    comment: review.comment || null,
-    observed_on: review.observed_on || null,
-    fruiting: review.fruiting ? review.fruiting.value : null,
-    quality_rating:
-      review.quality_rating === '0' ? null : Number(review.quality_rating) - 1,
-    yield_rating:
-      review.yield_rating === '0' ? null : Number(review.yield_rating) - 1,
-    photo_ids: review.photos.map((photo) => photo.id),
-  }
-  delete formattedReview.photos
-  return formattedReview
-}
-
-export const reviewToForm = ({
-  comment,
-  photos,
-  observed_on,
-  fruiting,
-  yield_rating,
-  quality_rating,
-}) => ({
-  comment: comment ?? '',
-  photos: photos.map((photo) => ({
-    id: photo.id,
-    name: `My Photo ${photo.created_at.split('T')[0]}`,
-    image: photo.thumb,
-    isNew: false,
-  })),
-  observed_on: observed_on ?? '',
-  fruiting: fruiting === null ? null : FRUITING_OPTIONS[fruiting],
-  yield_rating: yield_rating === null ? '0' : String(yield_rating + 1),
-  quality_rating: quality_rating === null ? '0' : String(quality_rating + 1),
-})
+import { FormWrapper } from './FormLayout'
 
 const DeleteButton = styled(Button)`
   background-color: ${({ theme }) => theme.invalid};
@@ -196,19 +88,6 @@ export const ReviewPhotoStep = () => (
   <PhotoUploader name="review.photos" label="Upload images" />
 )
 
-const StyledReviewForm = styled.div`
-  @media ${({ theme }) => theme.device.mobile} {
-    box-sizing: border-box;
-    width: 100%;
-    padding: 8px 27px 20px;
-    margin-top: 80px;
-
-    textarea {
-      height: 90px;
-    }
-  }
-`
-
 const RatingLabeledRow = styled(LabeledRow)`
   > div > label {
     color: ${({ theme }) => theme.text};
@@ -217,12 +96,10 @@ const RatingLabeledRow = styled(LabeledRow)`
 `
 
 export const ReviewForm = ({
-  stepped,
   initialValues = INITIAL_REVIEW_VALUES,
   editingId = null,
 }) => {
   const { locationId } = useSelector((state) => state.location)
-  const isLoggedIn = useSelector((state) => !!state.auth.user)
   const dispatch = useDispatch()
   const history = useAppHistory()
 
@@ -255,40 +132,32 @@ export const ReviewForm = ({
     }
   }
 
-  const { Recaptcha, handlePresubmit } = useInvisibleRecaptcha(handleSubmit)
-
-  const StepDisplay = stepped ? FormikStepper : FormikAllSteps
-
   return (
-    <StyledReviewForm>
-      <StepDisplay
-        onSubmit={isLoggedIn ? handleSubmit : handlePresubmit}
-        initialValues={initialValues}
-        validate={({ review }) => validateReview(review)}
-        renderButtons={({ isSubmitting, isValid }) => (
-          <ProgressButtons style={{ textAlign: editingId ? 'center' : 'left' }}>
-            <Button disabled={isSubmitting || !isValid} type="submit">
-              {isSubmitting ? 'Submitting' : 'Submit'}
-            </Button>
-            {editingId && (
-              <DeleteButton type="button" onClick={handleDelete}>
-                Delete
-              </DeleteButton>
-            )}
+    <FormWrapper
+      onSubmit={handleSubmit}
+      initialValues={initialValues}
+      validate={({ review }) => validateReview(review)}
+      renderButtons={(formikProps) => {
+        console.log(formikProps)
+        const { isSubmitting, isValid } = formikProps
+        return (
+          <ProgressButtons>
+            <div style={{ textAlign: editingId ? 'center' : 'left' }}>
+              <Button disabled={isSubmitting || !isValid} type="submit">
+                {isSubmitting ? 'Submitting' : 'Submit'}
+              </Button>
+              {editingId && (
+                <DeleteButton type="button" onClick={handleDelete}>
+                  Delete
+                </DeleteButton>
+              )}
+            </div>
           </ProgressButtons>
-        )}
-      >
-        <Step
-          label="Step 1"
-          validate={({ review }) => validateReviewStep(review)}
-        >
-          <ReviewStep standalone hasHeading={editingId == null && !stepped} />
-        </Step>
-        <Step label="Step 2" validate={({ review }) => validateReview(review)}>
-          <ReviewPhotoStep />
-        </Step>
-        {!isLoggedIn && <Recaptcha />}
-      </StepDisplay>
-    </StyledReviewForm>
+        )
+      }}
+    >
+      <ReviewStep standalone hasHeading={editingId == null} />
+      <ReviewPhotoStep />
+    </FormWrapper>
   )
 }
